pyproject.toml
[project]
name = "negotiator-purple"
version = "0.1.0"
description = "Purple Agent for MAizeBargAIn negotiation track (AgentX-AgentBeats)"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "pydantic>=2.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
